fix WebAttachmentSearchConfigClass static ClassFields
