+++ src/main/java/com/celements/search/web/classes/WebAttachmentSearchConfigClass.java
@@ -1,7 +1,5 @@
 package com.celements.search.web.classes;
 
-import static com.google.common.base.Preconditions.*;
-
 import java.util.List;
 
 import javax.annotation.concurrent.Immutable;
@@ -9,12 +7,10 @@ import javax.inject.Singleton;
 
 import org.xwiki.component.annotation.Component;
 
-import com.celements.marshalling.AbstractMarshaller;
 import com.celements.model.classes.AbstractClassDefinition;
 import com.celements.model.classes.fields.ClassField;
 import com.celements.model.classes.fields.list.CustomListField;
 import com.celements.model.classes.fields.list.StringListField;
-import com.google.common.base.Optional;
 import com.google.common.net.MediaType;
 
 @Immutable
@@ -26,13 +22,13 @@ public class WebAttachmentSearchConfigClass extends AbstractClassDefinition {
   public static final String DOC_NAME = "WebAttachmentSearchConfigClass";
   public static final String CLASS_DEF_HINT = SPACE_NAME + "." + DOC_NAME;
 
-  public ClassField<List<MediaType>> FIELD_MIMETYPES = new CustomListField.Builder<>(CLASS_DEF_HINT,
-      "mimeTypes", new MediaTypeMarshaller()).build();
+  public static ClassField<List<MediaType>> FIELD_MIMETYPES = new CustomListField.Builder<>(
+      CLASS_DEF_HINT, "mimeTypes", new MediaTypeMarshaller()).build();
 
-  public ClassField<List<MediaType>> FIELD_MIMETYPES_BLACK_LIST = new CustomListField.Builder<>(
+  public static ClassField<List<MediaType>> FIELD_MIMETYPES_BLACK_LIST = new CustomListField.Builder<>(
       CLASS_DEF_HINT, "mimeTypesBlackList", new MediaTypeMarshaller()).build();
 
-  public ClassField<List<String>> FIELD_FILENAME_PREFIXES = new StringListField.Builder(
+  public static ClassField<List<String>> FIELD_FILENAME_PREFIXES = new StringListField.Builder(
       CLASS_DEF_HINT, "fileNamePrefixes").build();
 
   @Override
@@ -53,31 +49,6 @@ public class WebAttachmentSearchConfigClass extends AbstractClassDefinition {
   @Override
   protected String getClassDocName() {
     return DOC_NAME;
-  }
-
-  @Immutable
-  private class MediaTypeMarshaller extends AbstractMarshaller<MediaType> {
-
-    public MediaTypeMarshaller() {
-      super(MediaType.class);
-    }
-
-    @Override
-    public String serialize(MediaType val) {
-      return val.toString();
-    }
-
-    @Override
-    public Optional<MediaType> resolve(String val) {
-      MediaType mediaType = null;
-      try {
-        mediaType = MediaType.parse(checkNotNull(val));
-      } catch (IllegalArgumentException exc) {
-        LOGGER.info("failed to resolve '{}' for '{}'", val, getToken(), exc);
-      }
-      return Optional.fromNullable(mediaType);
-    }
-
   }
 
 }
